backend/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-mysql-connector
